tests/unit/test_toc_discovery.py
"""
Unit tests for TOC discovery service.
"""

import pytest
from pathlib import Path
from PIL import Image
from app.services.toc_discovery import TOCDiscoveryService, MockTextract


class TestMockTextract:
    """Test MockTextract implementation."""
    
    def test_detect_document_text(self):
        """Test mock text detection."""
        mock = MockTextract()
        
        response = mock.detect_document_text(Document={'Bytes': b'fake'})
        
        assert 'Blocks' in response
        assert len(response['Blocks']) > 0
        assert response['Blocks'][0]['BlockType'] == 'LINE'


class TestTOCScoring:
    """Test TOC scoring algorithm."""
    
    @pytest.fixture
    def service(self):
        """Create TOC discovery service in local mode."""
        return TOCDiscoveryService(local_mode=True)
    
    def test_score_toc_with_keywords(self, service):
        """Test scoring with TOC keywords."""
        text = """
        Table of Contents
        
        Song Title 1 ............... 10
        Song Title 2 ............... 15
        Song Title 3 ............... 20
        """
        
        score = service.score_toc_likelihood(text)
        
        # Should have high score due to keyword + page numbers + dots
        assert score > 0.5
    
    def test_score_toc_with_page_numbers(self, service):
        """Test scoring with page numbers."""
        text = """
        Song Title 1 10
        Song Title 2 15
        Song Title 3 20
        Song Title 4 25
        Song Title 5 30
        """
        
        score = service.score_toc_likelihood(text)
        
        # Should have decent score due to page numbers
        assert score > 0.3
    
    def test_score_toc_with_dots(self, service):
        """Test scoring with leader dots."""
        text = """
        Song Title 1 ........ 10
        Song Title 2 ........ 15
        Song Title 3 ........ 20
        """
        
        score = service.score_toc_likelihood(text)
        
        # Should have good score due to dots + page numbers
        assert score > 0.4
    
    def test_score_non_toc_content(self, service):
        """Test scoring regular content page."""
        text = """
        This is a regular page with lots of text content.
        It has paragraphs and sentences that are much longer
        than typical TOC entries. There are no page numbers
        at the end of lines, and no leader dots connecting
        titles to page numbers.
        """
        
        score = service.score_toc_likelihood(text)
        
        # Should have low score (adjusted threshold based on actual scoring)
        assert score < 0.5
    
    def test_score_empty_text(self, service):
        """Test scoring empty text."""
        score = service.score_toc_likelihood("")
        assert score == 0.0
        
        score = service.score_toc_likelihood("   \n  \n  ")
        assert score == 0.0
    
    def test_score_short_text(self, service):
        """Test scoring very short text."""
        score = service.score_toc_likelihood("Short")
        assert score == 0.0
    
    def test_select_toc_pages_above_threshold(self, service):
        """Test selecting pages above threshold."""
        scored_pages = [
            (0, 0.8),
            (1, 0.9),
            (2, 0.3),
            (3, 0.6),
            (4, 0.2)
        ]
        
        selected = service.select_toc_pages(scored_pages, threshold=0.5)
        
        # Should select pages 0, 1, 3 (scores >= 0.5)
        assert set(selected) == {0, 1, 3}
        # Should be sorted by page number
        assert selected == [0, 1, 3]
    
    def test_select_toc_pages_none_above_threshold(self, service):
        """Test selecting pages when none meet threshold."""
        scored_pages = [
            (0, 0.3),
            (1, 0.2),
            (2, 0.4),
            (3, 0.1)
        ]
        
        selected = service.select_toc_pages(scored_pages, threshold=0.8)
        
        # Should select top 2 pages when none meet threshold
        assert len(selected) <= 2
        assert 2 in selected  # Highest score (0.4)
    
    def test_select_toc_pages_empty_list(self, service):
        """Test selecting from empty list."""
        selected = service.select_toc_pages([], threshold=0.5)
        assert selected == []


class TestTOCDiscoveryService:
    """Test TOC discovery service."""
    
    @pytest.fixture
    def service(self):
        """Create TOC discovery service in local mode."""
        return TOCDiscoveryService(local_mode=True)
    
    def test_initialization_local_mode(self):
        """Test service initialization in local mode."""
        service = TOCDiscoveryService(local_mode=True)
        assert service.local_mode is True
        assert isinstance(service.textract, MockTextract)
    
    def test_extract_text_from_image(self, service):
        """Test extracting text from image."""
        # Create a simple test image
        image = Image.new('RGB', (100, 100), color='white')
        
        text, confidence = service.extract_text_from_image(image, 0)
        
        # Mock should return some text
        assert isinstance(text, str)
        assert isinstance(confidence, float)
        assert 0.0 <= confidence <= 100.0
    
    def test_extract_text_textract(self, service):
        """Test extracting text from multiple images."""
        images = [
            Image.new('RGB', (100, 100), color='white'),
            Image.new('RGB', (100, 100), color='white')
        ]
        
        responses = service.extract_text_textract(images)
        
        assert len(responses) == 2
        assert all('page_num' in r for r in responses)
        assert all('text' in r for r in responses)
        assert all('confidence' in r for r in responses)


class TestTOCScoringHeuristics:
    """Test individual scoring heuristics."""
    
    @pytest.fixture
    def service(self):
        return TOCDiscoveryService(local_mode=True)
    
    def test_keyword_heuristic(self, service):
        """Test keyword detection."""
        # With keyword
        text_with_keyword = "Table of Contents\nSong 1\nSong 2"
        score_with = service.score_toc_likelihood(text_with_keyword)
        
        # Without keyword
        text_without = "Song 1\nSong 2\nSong 3"
        score_without = service.score_toc_likelihood(text_without)
        
        # Keyword should increase score
        assert score_with > score_without
    
    def test_page_number_heuristic(self, service):
        """Test page number detection."""
        # High ratio of page numbers
        text_high = "\n".join([f"Song {i} {i*5}" for i in range(1, 11)])
        score_high = service.score_toc_likelihood(text_high)
        
        # Low ratio of page numbers
        text_low = "\n".join([f"Song {i}" for i in range(1, 11)])
        score_low = service.score_toc_likelihood(text_low)
        
        # More page numbers should increase score (or at least not decrease it)
        assert score_high >= score_low
    
    def test_dots_heuristic(self, service):
        """Test leader dots detection."""
        # With dots
        text_with_dots = "\n".join([f"Song {i} ....... {i*5}" for i in range(1, 6)])
        score_with = service.score_toc_likelihood(text_with_dots)
        
        # Without dots
        text_without = "\n".join([f"Song {i} {i*5}" for i in range(1, 6)])
        score_without = service.score_toc_likelihood(text_without)
        
        # Dots should increase score
        assert score_with > score_without
    
    def test_combined_heuristics(self, service):
        """Test realistic TOC with multiple heuristics."""
        realistic_toc = """
        CONTENTS
        
        Amazing Grace .................. 5
        Ave Maria ...................... 12
        Canon in D ..................... 18
        Danny Boy ...................... 24
        Edelweiss ...................... 30
        """
        
        score = service.score_toc_likelihood(realistic_toc)
        
        # Should have very high score with multiple positive signals
        assert score > 0.7
