requirements.txt
# Core dependencies
boto3>=1.34.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
numpy>=1.24.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
hypothesis>=6.92.0

# Development dependencies
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0
